Add and fix code

+++ src/main/java/com/lvye/usercenterbackend/controller/UserController.java
@@ -6,11 +6,16 @@ import com.lvye.usercenterbackend.model.domain.request.userLoginRequest;
 import com.lvye.usercenterbackend.model.domain.request.userRegisterRequest;
 import com.lvye.usercenterbackend.service.UserService;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+
+import static com.lvye.usercenterbackend.constant.UserConstant.ADMIN_ROLE;
+import static com.lvye.usercenterbackend.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * @author LVye
@@ -51,7 +56,11 @@ public class UserController {
     }
 
     @GetMapping("/search")
-    public List<User> userSearch(String username) {
+    public List<User> userSearch(String username ,HttpServletRequest request) {
+        //是否管理员
+        if (!isAdmin(request)){
+            return new ArrayList<>();
+        }
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(username)) {
             queryWrapper.like("username", username);
@@ -59,11 +68,26 @@ public class UserController {
         return userService.list(queryWrapper);
     }
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody int id) {
+    public boolean deleteUser(@RequestBody long id,HttpServletRequest request) {
+        //是否管理员
+        if (!isAdmin(request)){
+            return false;
+        }
         if (id <= 0){
             return false;
         }
         return userService.removeById(id);
+    }
+
+    /**
+     *  判断是否为管理员
+     * @param request
+     * @return
+     */
+    private boolean isAdmin(HttpServletRequest request){
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user = (User) userObj;
+        return user != null && user.getUserRole() == ADMIN_ROLE;
     }
 }
